LayoutTests/fast/js/constructor-length.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This tests the length property of constructors.");

shouldBe('BeforeLoadEvent.length', '2');
shouldBe('CloseEvent.length', '2');
shouldBe('CustomEvent.length', '2');
shouldBe('DOMParser.length', '0');
shouldBe('DOMURL.length', '0');
shouldBe('ErrorEvent.length', '2');
shouldBe('Event.length', '2');
shouldBe('EventSource.length', '1');
shouldBe('FileReader.length', '0');
shouldBe('FileReaderSync.length', '0');
shouldBe('HashChangeEvent.length', '2');
shouldBe('Intent.length', '3');
shouldBe('MediaController.length', '0');
shouldBe('MediaStream.length', '2');
shouldBe('MessageChannel.length', '0');
shouldBe('MessageEvent.length', '2');
shouldBe('OverflowEvent.length', '2');
shouldBe('PageTransitionEvent.length', '2');
shouldBe('PeerConnection.length', '2');
shouldBe('PopStateEvent.length', '2');
shouldBe('ProgressEvent.length', '2');
shouldBe('ShadowRoot.length', '1');
shouldBe('SharedWorker.length', '2');
shouldBe('StorageEvent.length', '2');
shouldBe('TextTrackCue.length', '6');
shouldBe('TrackEvent.length', '2');
shouldBe('WebGLContextEvent.length', '2');
shouldBe('WebKitAnimationEvent.length', '2');
shouldBe('WebKitBlobBuilder.length', '0');
shouldBe('WebKitCSSMatrix.length', '1');
shouldBe('WebKitTransitionEvent.length', '2');
shouldBe('Worker.length', '1');
shouldBe('XMLHttpRequest.length', '0');
shouldBe('XMLSerializer.length', '0');
shouldBe('XPathEvaluator.length', '0');
shouldBe('XSLTProcessor.length', '0');

</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
